requirements.txt
aiogram==3.10.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.29
aiosqlite>=0.20.0
pydantic>=2.7.1
